Revert repository
