+++ auth-ui/src/main/java/com/google/android/horologist/auth/ui/googlesignin/prompt/GoogleSignInPromptViewModelFactory.kt
@@ -20,8 +20,6 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.ViewModelProvider.AndroidViewModelFactory.Companion.APPLICATION_KEY
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
-import com.google.android.gms.auth.api.signin.GoogleSignIn
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions
 import com.google.android.horologist.auth.data.googlesignin.GoogleSignInAuthUserRepository
 import com.google.android.horologist.auth.ui.ExperimentalHorologistAuthUiApi
 import com.google.android.horologist.auth.ui.common.screens.prompt.SignInPromptViewModel
@@ -35,13 +33,6 @@ public val GoogleSignInPromptViewModelFactory: ViewModelProvider.Factory = viewM
     initializer {
         val application = this[APPLICATION_KEY]!!
 
-        val googleSignInClient = GoogleSignIn.getClient(
-            application,
-            GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-                .requestEmail()
-                .build()
-        )
-
-        SignInPromptViewModel(GoogleSignInAuthUserRepository(application, googleSignInClient))
+        SignInPromptViewModel(GoogleSignInAuthUserRepository(application))
     }
 }
