+++ auth-ui/src/main/java/com/google/android/horologist/auth/ui/googlesignin/streamline/GoogleStreamlineSignInViewModelFactory.kt
@@ -19,8 +19,6 @@ package com.google.android.horologist.auth.ui.googlesignin.streamline
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
-import com.google.android.gms.auth.api.signin.GoogleSignIn
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions
 import com.google.android.horologist.auth.data.googlesignin.GoogleSignInAuthUserRepository
 import com.google.android.horologist.auth.ui.ExperimentalHorologistAuthUiApi
 import com.google.android.horologist.auth.ui.common.screens.streamline.StreamlineSignInViewModel
@@ -34,13 +32,6 @@ public val GoogleStreamlineSignInViewModelFactory: ViewModelProvider.Factory = v
     initializer {
         val application = this[ViewModelProvider.AndroidViewModelFactory.APPLICATION_KEY]!!
 
-        val googleSignInClient = GoogleSignIn.getClient(
-            application,
-            GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
-                .requestEmail()
-                .build()
-        )
-
-        StreamlineSignInViewModel(GoogleSignInAuthUserRepository(application, googleSignInClient))
+        StreamlineSignInViewModel(GoogleSignInAuthUserRepository(application))
     }
 }
